Restore double n inteasd of int n, where n = size of the storm

+++ boids.hpp
@@ -38,7 +38,7 @@ Point2D allineamento(std::vector<Boid> const& stormo, unsigned long int i,
 Point2D coesione(std::vector<Boid> const& stormo, unsigned long int i, float c,
                  float d, float angolo);
 
-std::vector<Boid> genera_stormo(int n);
+std::vector<Boid> genera_stormo(double n);
 
 void movimento(std::vector<Boid>& stormo, float t);
